refactoring, examples update
