+++ catalyst/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '19.02-rc0'
+__version__ = '19.03-rc0'
