+++ examples/cifar_simple/__init__.py
@@ -1,0 +1,3 @@
+from catalyst.dl.experiments.runner import SupervisedRunner as Runner
+from .experiment import Experiment
+from .model import SimpleNet as Model
