+++ examples/cifar_simple/model.py
@@ -1,6 +1,5 @@
 import torch.nn as nn
 import torch.nn.functional as F
-from catalyst.dl.runner import SupervisedModelRunner as ModelRunner  # noqa F401
 from catalyst.contrib.registry import Registry
