+++ examples/cifar_stages/README.md
@@ -3,7 +3,7 @@
 ### Local run
 
 ```bash
-catalyst-dl train --config=./cifar_stages/config.yml
+catalyst-dl run --config=./cifar_stages/config.yml
 ```
 
 ### Docker run
@@ -17,7 +17,7 @@ docker run -it --rm --runtime=nvidia \
    -e "CUDA_VISIBLE_DEVICES=0" \
    -e "LOGDIR=/logdir" \
    catalyst-base \
-   catalyst-dl train --config=./cifar_stages/config.yml --logdir=/logdir
+   catalyst-dl run --config=./cifar_stages/config.yml --logdir=/logdir
 ```
