--- examples/finetune_new/config.yml
@@ -1,10 +1,0 @@
-stages:
-  stage_1:
-    epochs: 10
-    data_params:
-      batch_size: 100
-
-
-
-
-
